+++ packages/guide/src/app/npcs/components/gifting/gifting.component.ts
@@ -1,9 +1,14 @@
 import { Component, inject, ViewEncapsulation } from '@angular/core';
 import { DatabaseService } from '../../../shared/services/database.service';
-import { Observable } from 'rxjs';
-import { GiftPreferences } from '@ci/data-types';
+import { forkJoin, Observable, of, switchMap } from 'rxjs';
+import { GiftPreferences, NPC } from '@ci/data-types';
 import { MapKeyed } from '../../../shared/types/map-keyed.type';
 import { UiIcon } from '../../../shared/enums/ui-icon.enum';
+
+type CombinedGiftPreference = {
+    preferences: MapKeyed<GiftPreferences>,
+    npc: NPC | undefined
+}
 
 @Component({
     selector: 'app-gifting',
@@ -13,9 +18,11 @@ import { UiIcon } from '../../../shared/enums/ui-icon.enum';
 })
 export class GiftingComponent {
 
-    gifting$: Observable<MapKeyed<GiftPreferences>[]> = inject(DatabaseService).fetchGiftingPreferences$();
-
-    preferencesMap: { icon: UiIcon, label: string; preferenceField: string }[] = [
+    preferencesMap: {
+        icon: UiIcon,
+        label: string;
+        preferenceField: string
+    }[] = [
         {icon: UiIcon.LOVE, label: 'Favorite', preferenceField: 'favoritePreferences'},
         {icon: UiIcon.LOVE, label: 'Love', preferenceField: 'lovePreferences'},
         {icon: UiIcon.LIKE, label: 'Like', preferenceField: 'likePreferences'},
@@ -23,6 +30,27 @@ export class GiftingComponent {
         {icon: UiIcon.DISLIKE, label: 'Dislike', preferenceField: 'dislikePreferences'},
         {icon: UiIcon.HATE, label: 'Hate', preferenceField: 'hatePreferences'},
     ];
+    protected uiIcon = UiIcon;
+    protected gifting$: Observable<CombinedGiftPreference[]>;
+    private _database = inject(DatabaseService)
+
+    constructor() {
+        this.gifting$ = forkJoin({
+            gifts: this._database.fetchGiftingPreferences$(),
+            npcs: this._database.fetchNPCs$(),
+        }).pipe(
+            switchMap(({gifts, npcs}) => {
+
+                const mappedGifts: CombinedGiftPreference[] = gifts.map(gift => {
+                    return {
+                        preferences: gift,
+                        npc: npcs.find(npc => npc.key === gift.mapKey)
+                    } satisfies CombinedGiftPreference
+                })
+                return of(mappedGifts)
+            })
+        )
+    }
 
 
 }