+++ packages/guide/CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added merit exchange
 - Added PWA support
 - Added manual mapping for processors to show in database to reduce confusion about internal names
+- Added birthdays to NPC gifting
 
 #### Changed